iapplx/rpmbuild: Skip the Babel step for now
This makes the minimum supported BIG-IP version 14.1 for the time being.
While the Babel version of the code runs (the worker loads and end
points can be hit), template loading is buggy.

This will need to be explored (probably try running Babel version
through unit tests) so we can re-enable 13.1 support.