iapplx/worker: Add install installtemplateset endpoint to POST to
There is a name parameter for the name of the zip file (without
extension) to install as a template set. This assumes the template set
files are in the root of the zip file (i.e.., there is not a directory
containing the template set files in the zip file).