iappslx/gui: Update json-editor to a patched 2.2.0 preview build
This build is based on fix-703 branch of the fork at
https://github.com/mstokes-f5/json-editor/. This branch includes a patch
to deal with editors that have multiple dependencies. The patch has been
submitted upstream as a  PR
(https://github.com/json-editor/json-editor/pull/704), and a bug has
been filed as (https://github.com/json-editor/json-editor/issues/703).

If/when the PR is merged, we can abandon the fork.