iappslx/worker: Add the start of integration tests
This can be run with "npm run testinteg" and requires BIGIP_TARGET and
BIGIP_CREDS environment variables to be set. This integration test will
try deploying a couple of applications to the target BIG-IP using the
FAST REST API. Onboarding the box and installing FAST and AS3 needs to
be done before running the integration test (i.e., should be done in a
separate, earlier pipeline stage).