worker: Support bigipDependencies property on templates
This will fail to load a template that depends on module that is not
provisioned. Sub-templates in an anyOf that fail the dependency check
will be silently stripped from the root template. For oneOf, if at least
one item remains in the array after stripping sub-templates, the root
template will still load successfully. Any failed sub-template in an
allOf will result in a failure to load the root template.