AS3 Driver: Use AS3's optimistic lock feature
This will report an error if FAST is out-of-sync with AS3. Since FAST
currently uses AS3 as a source-of-truth, this is unlikely to happen
unless:

1. There are other AS3 operations happening during FAST's
read-modify-write cycle. Currently AS3 only allows a single operation at
a time, so the window for this is narrow.
2. FAST's AS3 cache has gotten stale and was not properly invalidated.
This would be a FAST bug that needs to be fixed, and not throwing an
error could lead to worse problems.

An option still needs to be added to allow disabling this lock.