worker: Add /render endpoint
A POST to this endpoint works like a POST to /applications except the
rendered template result is returned (in a list) instead of forwarded to
AS3.

NOTE: This returned AS3 declaration is not stitched with other
applications on box so submitting it to AS3 directly can result in data
loss.