gui: Attempt to reuse auth token from the BIG-IP GUI
This prevents additional basic auth login prompts as the GUI attempts to
access the BIG-IP REST API.

Getting the data (digest and timenow) to get the auth token from AuthN
is pretty hacky. Most of restui/xui seem to assume this value is just
passed around as form inputs or URL query parameters. However, this does
not appear to get passed along to custom iApps LX presentation layers.
Instead we peek at the parent frame for links known to have these values
as query parameters.

This GUI data appears to be reliably available in 13.1, 15.1, and 16.0.

closes MYSTIQUE-397
closes MYSTIQUE-404