gui: Do not overwrite existing applications when creating new ones
A new "allowOverwrite" property has been added to the POST body for the
/applications endpoint. This defaults to true, but if set to false will
prevent updating the AS3 declaration if the rendered application result
would overwrite an existing application in the declaration.

The GUI sets this property based on whether it is creating a new
application or modifying an existing one.

Closes MYSTIQUE-407